upgrade windows dependencies and make driver_com.rs more modular
