+++ owlyshield_predict/src/driver_com/README.md
@@ -1,0 +1,5 @@
+- [mod.rs](mod.rs) contains Low-level communication with the minifilter.
+
+- [shared_def.rs](shared_def.rs) contains all definitions shared between this user-mode app and the minifilter in order
+  to communicate properly. Those are C-representation of structures sent or received from the minifilter.
+
