+++ owlyshield_predict/src/notifications.rs
@@ -46,7 +46,7 @@ pub fn toast(config: &Config, message: &str, report_path: &str) -> Result<(), St
             if !DuplicateTokenEx(
                 service_token,
                 TOKEN_ALL_ACCESS,
-                null_mut() as *mut SECURITY_ATTRIBUTES,
+                None,
                 SecurityIdentification,
                 TokenPrimary,
                 &mut token,
@@ -63,11 +63,11 @@ pub fn toast(config: &Config, message: &str, report_path: &str) -> Result<(), St
                 token,
                 PCWSTR(str_to_pcwstr(toastapp_path.to_str().unwrap()).into_raw()),
                 PWSTR(str_to_pwstr(&toastapp_args).into_vec().as_mut_ptr()),
-                null_mut(),
-                null_mut(),
+                None,
+                None,
                 BOOL(0),
                 CREATE_NEW_CONSOLE.0,
-                null_mut(),
+                None,
                 PCWSTR(str_to_pcwstr(toastapp_dir.to_str().unwrap()).into_raw()),
                 std::ptr::addr_of_mut!(si),
                 std::ptr::addr_of_mut!(pi),
